+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Update format used for durations in `rics.performance`.
+
 ## [3.0.1] - 2023-03-25
 
 ### Changed